Adds actions to reflect the number of ways that cards can produce effects
In Agricola, there is a technical distinction between taking a named action, such as 'Plow 1 field', and gaining the effect of being able to plow 1 field. This mirrors the distinction in acquiring goods between 'taking' them off the board, and 'receiving' them from the general supply as the result of a card effect.